requirements.txt
aiohappyeyeballs==2.6.1
asgiref==3.9.1
Django==5.1.3
django-filter==25.1
djangorestframework==3.16.1
Markdown==3.8.2
sqlparse==0.5.3
environs~=14.3.0
pillow~=11.3.0
djangorestframework_simplejwt==5.5.1
channels==4.3.1
channels_redis==4.3.0
django-cors-headers==4.7.0
drf-yasg==1.21.10
inflection==0.5.1
marshmallow==4.0.0
msgpack==1.1.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
uritemplate==4.2.0
firebase_admin~=7.1.0
